+++ feature/src/main/java/com/terning/feature/filtering/filteringthree/FilteringThreeSideEffect.kt
@@ -1,4 +1,4 @@
-package com.terning.feature.filtering.filtering
+package com.terning.feature.filtering.filteringthree
 
 import androidx.annotation.StringRes
