+++ feature/src/main/java/com/terning/feature/filtering/filteringthree/FilteringViewModel.kt
@@ -1,4 +1,4 @@
-package com.terning.feature.filtering.filtering
+package com.terning.feature.filtering.filteringthree
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
